Implement a MovieType builder class
