+++ builderClass/Builders/MovieTypeBuilder.cpp
@@ -1,0 +1,18 @@
+#include "MovieTypeBuilder.h"
+
+MovieTypeBuilder::MovieTypeBuilder(){
+
+}
+
+MovieTypeBuilder::~MovieTypeBuilder(){
+
+}
+
+MovieType MovieTypeBuilder::build(string& title, string& synopsis, double& rating){
+    MovieType m;
+    m.title = title;
+    m.synopsis = synopsis;
+    m.rating = rating;
+    return m;
+}
+
